Get the feed for a location

This function filters by location and returns all posts that have the same location

ig_get_location_feed(location_id, max_id = NULL, ranked_content = TRUE,
  return_df = TRUE, paginate = TRUE, max_pages = 10,
  verbose = FALSE)

Arguments

location_id
integer; the unique id to identify a place which can be found in the "pk", "external_id", or "facebook_places_id" fields on other objects returned via the Instagram API

max_id
integer; the unique id identifying the oldest post that you would want to retrieve in this function call

ranked_content
logical; do you want the feed content to be sorted by rank?

return_df
logical; do you want to return the results as a tbl_df with one row per entity or as a list with one element per entity?

paginate
logical; do you want to paginate through results or just return the first page?

max_pages
integer; a limit to the number of pages to retrieve from paginated endpoints. Instagram feeds have the potential to paginate forever, so by default we stop after pulling 10 pages. If you would like more or less pages returned, then modify this argument.

verbose
logical; do you want informative messages?

Details

Note that if your location is a "group" (such as a city), the feed will include media from multiple locations within that area. But if your location is a very specific place such as a specific night club, it will usually only include media from that exact location.

See also

Examples

# NOT RUN {
# location feed for the city of Washington, DC
ig_get_location_feed(213480180)
# }