Search by Hashtag

This function accepts a hashtag, without the "#" symbol and returns the similar hashtags. You can then use ig_get_hashtag_feed() to return all posts associated with that hashtag.

ig_search_tags(hashtag, max_id = NULL, return_df = TRUE,
  paginate = TRUE, max_pages = 10, verbose = FALSE)

Arguments

hashtag
character; do not include the hashtag (pound sign) at the beginning of the string

max_id
integer; the unique id identifying the oldest post that you would want to retrieve in this function call

return_df
logical; do you want to return the results as a tbl_df with one row per entity or as a list with one element per entity?

paginate
logical; do you want to paginate through results or just return the first page?

max_pages
integer; a limit to the number of pages to retrieve from paginated endpoints. Instagram feeds have the potential to paginate forever, so by default we stop after pulling 10 pages. If you would like more or less pages returned, then modify this argument.

verbose
logical; do you want informative messages?

Examples

# NOT RUN {
# search for posts using the #belieber hashtag
belieber_posts <- ig_search_tags("belieber")
# }